[Refactor] 홈 뷰컨 추가 (#9)
