+++ Tving_CloneProject/Tving_CloneProject/Global/Components/TabBarViewController.swift
@@ -13,7 +13,7 @@ final class TabBarViewController: UITabBarController {
     
     // MARK: - UI Components
     
-    let mainVC = MainViewController()
+    let mainVC = HomeViewController()
     
     let toBeReleasedVC = ToBeReleasedViewController()
